Mouse codes and move of keyCodes to package input
