+++ API/Input/mouseCodes.go
@@ -1,0 +1,22 @@
+package Input
+
+import "C"
+
+// MouseButton corresponds to a mouse button.
+type MouseButton int
+
+// Mouse buttons.
+const (
+	MouseButton1      MouseButton = 0
+	MouseButton2      MouseButton = 1
+	MouseButton3      MouseButton = 2
+	MouseButton4      MouseButton = 3
+	MouseButton5      MouseButton = 4
+	MouseButton6      MouseButton = 5
+	MouseButton7      MouseButton = 6
+	MouseButton8      MouseButton = 7
+	MouseButtonLast               = MouseButton8
+	MouseButtonLeft               = MouseButton1
+	MouseButtonRight              = MouseButton2
+	MouseButtonMiddle             = MouseButton3
+)
